requirements.txt
streamlit
pandas
plotly
Pillow
scipy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
